Extract package parsing logic to static method
